Refine core data
