+++ nautilus_core/model/src/data/trade.rs
@@ -1,0 +1,110 @@
+// -------------------------------------------------------------------------------------------------
+//  Copyright (C) 2015-2023 Nautech Systems Pty Ltd. All rights reserved.
+//  https://nautechsystems.io
+//
+//  Licensed under the GNU Lesser General Public License Version 3.0 (the "License");
+//  You may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at https://www.gnu.org/licenses/lgpl-3.0.en.html
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+// -------------------------------------------------------------------------------------------------
+
+use std::fmt::{Display, Formatter};
+
+use nautilus_core::time::UnixNanos;
+use pyo3::prelude::*;
+use serde::{Deserialize, Serialize};
+
+use crate::{
+    enums::AggressorSide,
+    identifiers::{instrument_id::InstrumentId, trade_id::TradeId},
+    types::{price::Price, quantity::Quantity},
+};
+
+/// Represents a single trade tick in a financial market.
+#[repr(C)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[pyclass]
+pub struct TradeTick {
+    pub instrument_id: InstrumentId,
+    pub price: Price,
+    pub size: Quantity,
+    pub aggressor_side: AggressorSide,
+    pub trade_id: TradeId,
+    pub ts_event: UnixNanos,
+    pub ts_init: UnixNanos,
+}
+
+impl TradeTick {
+    #[must_use]
+    pub fn new(
+        instrument_id: InstrumentId,
+        price: Price,
+        size: Quantity,
+        aggressor_side: AggressorSide,
+        trade_id: TradeId,
+        ts_event: UnixNanos,
+        ts_init: UnixNanos,
+    ) -> Self {
+        Self {
+            instrument_id,
+            price,
+            size,
+            aggressor_side,
+            trade_id,
+            ts_event,
+            ts_init,
+        }
+    }
+}
+
+impl Display for TradeTick {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{},{},{},{},{},{}",
+            self.instrument_id,
+            self.price,
+            self.size,
+            self.aggressor_side,
+            self.trade_id,
+            self.ts_event,
+        )
+    }
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Tests
+////////////////////////////////////////////////////////////////////////////////
+#[cfg(test)]
+mod tests {
+    use std::str::FromStr;
+
+    use crate::{
+        data::trade::TradeTick,
+        enums::AggressorSide,
+        identifiers::{instrument_id::InstrumentId, trade_id::TradeId},
+        types::{price::Price, quantity::Quantity},
+    };
+
+    #[test]
+    fn test_trade_tick_to_string() {
+        let tick = TradeTick {
+            instrument_id: InstrumentId::from_str("ETHUSDT-PERP.BINANCE").unwrap(),
+            price: Price::new(10000.0, 4),
+            size: Quantity::new(1.0, 8),
+            aggressor_side: AggressorSide::Buyer,
+            trade_id: TradeId::new("123456789"),
+            ts_event: 0,
+            ts_init: 0,
+        };
+        assert_eq!(
+            tick.to_string(),
+            "ETHUSDT-PERP.BINANCE,10000.0000,1.00000000,BUYER,123456789,0"
+        );
+    }
+}
